JsonDocument god class divided into logical units
